Simplify unit tests
