+++ src/SourceLink.Tfs.Git.UnitTests/GetSourceLinkUrlTests.cs
@@ -29,407 +29,11 @@ namespace Microsoft.SourceLink.Tfs.Git.UnitTests
         }
 
         [Theory]
-        [InlineData("mytfs*.com")]
-        [InlineData("mytfs.com/a")]
-        [InlineData("mytfs.com/a?x=2")]
-        [InlineData("http://mytfs.com")]
-        [InlineData("http://a@mytfs.com")]
-        [InlineData("a@mytfs.com")]
-        public void HostsDomain_Errors(string domain)
-        {
-            var engine = new MockEngine();
-
-            var task = new GetSourceLinkUrl()
-            {
-                BuildEngine = engine,
-                SourceRoot = new MockItem("x", KVP("RepositoryUrl", "http://abc.com"), KVP("SourceControl", "git")),
-                Hosts = new[] { new MockItem(domain) }
-            };
-
-            bool result = task.Execute();
-
-            AssertEx.AssertEqualToleratingWhitespaceDifferences(
-                "ERROR : " + string.Format(CommonResources.ValuePassedToTaskParameterNotValidDomainName, "Hosts", domain), engine.Log);
-
-            Assert.False(result);
-        }
-
-        [Theory]
-        [InlineData("http://mytfs*.com")]
-        // TODO (https://github.com/dotnet/sourcelink/issues/120): fails on Linux [InlineData("/a")]
-        public void ImplicitHost_Errors(string repositoryUrl)
-        {
-            var engine = new MockEngine();
-
-            var task = new GetSourceLinkUrl()
-            {
-                BuildEngine = engine,
-                SourceRoot = new MockItem("x", KVP("RepositoryUrl", "http://abc.com"), KVP("SourceControl", "git")),
-                RepositoryUrl = repositoryUrl,
-                IsSingleProvider = true,
-            };
-
-            bool result = task.Execute();
-
-            AssertEx.AssertEqualToleratingWhitespaceDifferences(
-                "ERROR : " + string.Format(CommonResources.ValuePassedToTaskParameterNotValidUri, "RepositoryUrl", repositoryUrl), engine.Log);
-
-            Assert.False(result);
-        }
-
-        [Theory]
-        [InlineData("mytfs.com")]
-        [InlineData("mytfs.com/a")]
-        [InlineData("mytfs.com/a?x=2")]
-        [InlineData("http://a@mytfs.com")]
-        [InlineData("a@mytfs.com")]
-        public void HostsContentUrl_Errors(string url)
-        {
-            var engine = new MockEngine();
-
-            var task = new GetSourceLinkUrl()
-            {
-                BuildEngine = engine,
-                SourceRoot = new MockItem("x", KVP("RepositoryUrl", "http://abc.com"), KVP("SourceControl", "git")),
-                Hosts = new[] { new MockItem("abc.com", KVP("ContentUrl", url)) }
-            };
-
-            bool result = task.Execute();
-
-            AssertEx.AssertEqualToleratingWhitespaceDifferences(
-                "ERROR : " + string.Format(CommonResources.ValuePassedToTaskParameterNotValidHostUri, "Hosts", url), engine.Log);
-
-            Assert.False(result);
-        }
-
-        [Theory]
-        [InlineData("a/b")]
-        [InlineData("")]
-        [InlineData("http://")]
-        public void RepositoryUrl_Errors(string url)
-        {
-            var engine = new MockEngine();
-
-            var task = new GetSourceLinkUrl()
-            {
-                BuildEngine = engine,
-                SourceRoot = new MockItem("/src/", KVP("RepositoryUrl", url), KVP("SourceControl", "git")),
-            };
-
-            bool result = task.Execute();
-
-            AssertEx.AssertEqualToleratingWhitespaceDifferences(
-                "ERROR : " + string.Format(CommonResources.ValueOfWithIdentityIsInvalid, "SourceRoot.RepositoryUrl", "/src/", url), engine.Log);
-
-            Assert.False(result);
-        }
-
-        [Theory]
-        [InlineData("123")]
-        [InlineData(" 00000000000000000000000000000000000000")]
-        [InlineData("000000000000000000000000000000000000000G")]
-        [InlineData("000000000000000000000000000000000000000g")]
-        [InlineData("00000000000000000000000000000000000000001")]
-        [InlineData("")]
-        public void RevisionId_Errors(string revisionId)
-        {
-            var engine = new MockEngine();
-
-            var task = new GetSourceLinkUrl()
-            {
-                BuildEngine = engine,
-                SourceRoot = new MockItem("/src/", KVP("RepositoryUrl", "http://tfs.com/a/b"), KVP("SourceControl", "git"), KVP("RevisionId", revisionId)),
-                Hosts = new[] { new MockItem("tfs.com", KVP("ContentUrl", "https://tfs.com")) }
-            };
-
-            bool result = task.Execute();
-
-            AssertEx.AssertEqualToleratingWhitespaceDifferences(
-                "ERROR : " + string.Format(CommonResources.ValueOfWithIdentityIsNotValidCommitHash, "SourceRoot.RevisionId", "/src/", revisionId), engine.Log);
-
-            Assert.False(result);
-        }
-
-        [Fact]
-        public void SourceRootNotApplicable_SourceControlNotGit()
-        {
-            var engine = new MockEngine();
-
-            var task = new GetSourceLinkUrl()
-            {
-                BuildEngine = engine,
-                SourceRoot = new MockItem("/src/", KVP("RepositoryUrl", "http://tfs.com/a/b"), KVP("SourceControl", "tfvc"), KVP("RevisionId", "12345")),
-                Hosts = new[] { new MockItem("tfs.com", KVP("ContentUrl", "https://tfs.com")) }
-            };
-
-            bool result = task.Execute();
-            AssertEx.AssertEqualToleratingWhitespaceDifferences("", engine.Log);
-            Assert.Equal("N/A", task.SourceLinkUrl);
-            Assert.True(result);
-        }
-
-        [Fact]
-        public void SourceRootNotApplicable_SourceLinkUrlSet()
-        {
-            var engine = new MockEngine();
-
-            var task = new GetSourceLinkUrl()
-            {
-                BuildEngine = engine,
-                SourceRoot = new MockItem("/src/", KVP("RepositoryUrl", "http://tfs.com/a/b"), KVP("SourceControl", "git"), KVP("SourceLinkUrl", "x"), KVP("RevisionId", "12345")),
-                Hosts = new[] { new MockItem("github.com", KVP("ContentUrl", "https://tfs.com")) }
-            };
-
-            bool result = task.Execute();
-            AssertEx.AssertEqualToleratingWhitespaceDifferences("", engine.Log);
-            Assert.Equal("N/A", task.SourceLinkUrl);
-            Assert.True(result);
-        }
-
-        [Fact]
-        public void SourceRootNotApplicable_RepositoryUrlNotMatchingHost()
-        {
-            var engine = new MockEngine();
-
-            var task = new GetSourceLinkUrl()
-            {
-                BuildEngine = engine,
-                SourceRoot = new MockItem("/src/", KVP("RepositoryUrl", "http://abc.com/a/b"), KVP("SourceControl", "git"), KVP("RevisionId", "12345")),
-                Hosts = new[]
-                {
-                    new MockItem("mytfs1.com", KVP("ContentUrl", "https://mytfs1.com")),
-                    new MockItem("mytfs2.com", KVP("ContentUrl", "http://mytfs2.com"))
-                }
-            };
-
-            bool result = task.Execute();
-            AssertEx.AssertEqualToleratingWhitespaceDifferences("", engine.Log);
-            Assert.Equal("N/A", task.SourceLinkUrl);
-            Assert.True(result);
-        }
-
-        [Fact]
-        public void CustomHosts_PortWithDefaultContentUrl()
-        {
-            var engine = new MockEngine();
-
-            var task = new GetSourceLinkUrl()
-            {
-                BuildEngine = engine,
-                SourceRoot = new MockItem("/src/", KVP("RepositoryUrl", "http://subdomain.tfs.com:1234/collection/project/_git/repo"), KVP("SourceControl", "git"), KVP("RevisionId", "0123456789abcdefABCDEF000000000000000000")),
-                Hosts = new[]
-                {
-                    new MockItem("tfs.com"),
-                }
-            };
-
-            bool result = task.Execute();
-            AssertEx.AssertEqualToleratingWhitespaceDifferences("", engine.Log);
-            AssertEx.AreEqual("https://tfs.com:1234/collection/project/_apis/git/repositories/repo/items?api-version=1.0&versionType=commit&version=0123456789abcdefABCDEF000000000000000000&path=/*", task.SourceLinkUrl);
-            Assert.True(result);
-        }
-
-        [Fact]
-        public void ImplicitHost_PortWithDefaultContentUrl()
-        {
-            var engine = new MockEngine();
-
-            var task = new GetSourceLinkUrl()
-            {
-                BuildEngine = engine,
-                SourceRoot = new MockItem("/src/", KVP("RepositoryUrl", "http://subdomain.tfs.com:1234/collection/project/_git/repo"), KVP("SourceControl", "git"), KVP("RevisionId", "0123456789abcdefABCDEF000000000000000000")),
-                RepositoryUrl = "http://tfs.com",
-                IsSingleProvider = true,
-            };
-
-            bool result = task.Execute();
-            AssertEx.AssertEqualToleratingWhitespaceDifferences("", engine.Log);
-            AssertEx.AreEqual("https://tfs.com:1234/collection/project/_apis/git/repositories/repo/items?api-version=1.0&versionType=commit&version=0123456789abcdefABCDEF000000000000000000&path=/*", task.SourceLinkUrl);
-            Assert.True(result);
-        }
-
-        [Fact]
-        public void CustomHosts_PortWithNonDefaultContentUrl()
-        {
-            var engine = new MockEngine();
-
-            var task = new GetSourceLinkUrl()
-            {
-                BuildEngine = engine,
-                SourceRoot = new MockItem("/src/", KVP("RepositoryUrl", "http://subdomain.tfs.com:1234/collection/project/_git/repo"), KVP("SourceControl", "git"), KVP("RevisionId", "0123456789abcdefABCDEF000000000000000000")),
-                Hosts = new[]
-                {
-                    new MockItem("tfs.com", KVP("ContentUrl", "https://othertfs.com")),
-                }
-            };
-
-            bool result = task.Execute();
-            AssertEx.AssertEqualToleratingWhitespaceDifferences("", engine.Log);
-            AssertEx.AreEqual("https://othertfs.com/collection/project/_apis/git/repositories/repo/items?api-version=1.0&versionType=commit&version=0123456789abcdefABCDEF000000000000000000&path=/*", task.SourceLinkUrl);
-            Assert.True(result);
-        }
-
-        [Fact]
-        public void CustomHosts_Matching1()
-        {
-            var engine = new MockEngine();
-
-            var task = new GetSourceLinkUrl()
-            {
-                BuildEngine = engine,
-                SourceRoot = new MockItem("/src/", KVP("RepositoryUrl", "http://subdomain.tfs.com:1234/collection/project/_git/repo"), KVP("SourceControl", "git"), KVP("RevisionId", "0123456789abcdefABCDEF000000000000000000")),
-                RepositoryUrl = "http://abc.com",
-                IsSingleProvider = true,
-                Hosts = new[]
-                {
-                    new MockItem("contoso.com", KVP("ContentUrl", "https://contoso.com")),
-                    new MockItem("tfs.com", KVP("ContentUrl", "https://subdomain.tfs.com1:777")),
-                    new MockItem("tfs.com", KVP("ContentUrl", "https://subdomain.tfs.com2"))
-                }
-            };
-
-            bool result = task.Execute();
-            AssertEx.AssertEqualToleratingWhitespaceDifferences("", engine.Log);
-            AssertEx.AreEqual("https://subdomain.tfs.com1:777/collection/project/_apis/git/repositories/repo/items?api-version=1.0&versionType=commit&version=0123456789abcdefABCDEF000000000000000000&path=/*", task.SourceLinkUrl);
-            Assert.True(result);
-        }
-
-        [Fact]
-        public void CustomHosts_Matching2()
-        {
-            var engine = new MockEngine();
-
-            var task = new GetSourceLinkUrl()
-            {
-                BuildEngine = engine,
-                SourceRoot = new MockItem("/src/", KVP("RepositoryUrl", "http://subdomain.tfs.com:123/collection/project/_git/repo"), KVP("SourceControl", "git"), KVP("RevisionId", "0123456789abcdefABCDEF000000000000000000")),
-                RepositoryUrl = "http://subdomain.tfs.com:123",
-                IsSingleProvider = true,
-                Hosts = new[]
-                {
-                    new MockItem("tfs.com", KVP("ContentUrl", "https://domain.com:1")),
-                    new MockItem("tfs.com:123", KVP("ContentUrl", "https://domain.com:2")),
-                    new MockItem("tfs.com:123", KVP("ContentUrl", "https://domain.com:3")),
-                    new MockItem("subdomain.tfs.com", KVP("ContentUrl", "https://domain.com:4")),
-                    new MockItem("subdomain.tfs.com:123", KVP("ContentUrl", "https://domain.com:5")),
-                }
-            };
-
-            bool result = task.Execute();
-            AssertEx.AssertEqualToleratingWhitespaceDifferences("", engine.Log);
-
-            // explicit host is preferred over the implicit one 
-            AssertEx.AreEqual("https://domain.com:5/collection/project/_apis/git/repositories/repo/items?api-version=1.0&versionType=commit&version=0123456789abcdefABCDEF000000000000000000&path=/*", task.SourceLinkUrl);
-            Assert.True(result);
-        }
-
-        [Fact]
-        public void CustomHosts_Matching3()
-        {
-            var engine = new MockEngine();
-
-            var task = new GetSourceLinkUrl()
-            {
-                BuildEngine = engine,
-                SourceRoot = new MockItem("/src/", KVP("RepositoryUrl", "http://subdomain.tfs.com:100/collection/project/_git/repo"), KVP("SourceControl", "git"), KVP("RevisionId", "0123456789abcdefABCDEF000000000000000000")),
-                Hosts = new[]
-                {
-                    new MockItem("tfs.com", KVP("ContentUrl", "https://domain.com:1")),
-                    new MockItem("tfs.com:123", KVP("ContentUrl", "https://domain.com:2")),
-                    new MockItem("tfs.com:123", KVP("ContentUrl", "https://domain.com:3")),
-                    new MockItem("subdomain.tfs.com", KVP("ContentUrl", "https://domain.com:4")),
-                    new MockItem("subdomain.tfs.com:123", KVP("ContentUrl", "https://domain.com:5")),
-                    new MockItem("subdomain.tfs.com:123", KVP("ContentUrl", "https://domain.com:6")),
-                }
-            };
-
-            bool result = task.Execute();
-            AssertEx.AssertEqualToleratingWhitespaceDifferences("", engine.Log);
-            AssertEx.AreEqual("https://domain.com:4/collection/project/_apis/git/repositories/repo/items?api-version=1.0&versionType=commit&version=0123456789abcdefABCDEF000000000000000000&path=/*", task.SourceLinkUrl);
-            Assert.True(result);
-        }
-
-        [Fact]
-        public void CustomHosts_Matching4()
-        {
-            var engine = new MockEngine();
-
-            var task = new GetSourceLinkUrl()
-            {
-                BuildEngine = engine,
-                SourceRoot = new MockItem("/src/", KVP("RepositoryUrl", "http://subdomain.tfs.com:123/collection/project/_git/repo"), KVP("SourceControl", "git"), KVP("RevisionId", "0123456789abcdefABCDEF000000000000000000")),
-                Hosts = new[]
-                {
-                    new MockItem("tfs.com", KVP("ContentUrl", "https://domain.com:1")),
-                    new MockItem("tfs.com:123", KVP("ContentUrl", "https://domain.com:2")),
-                    new MockItem("tfs.com:123", KVP("ContentUrl", "https://domain.com:3")),
-                    new MockItem("z.tfs.com", KVP("ContentUrl", "https://domain.com:4")),
-                    new MockItem("z.tfs.com:123", KVP("ContentUrl", "https://domain.com:5")),
-                    new MockItem("z.tfs.com:123", KVP("ContentUrl", "https://domain.com:6")),
-                }
-            };
-
-            bool result = task.Execute();
-            AssertEx.AssertEqualToleratingWhitespaceDifferences("", engine.Log);
-            AssertEx.AreEqual("https://domain.com:2/collection/project/_apis/git/repositories/repo/items?api-version=1.0&versionType=commit&version=0123456789abcdefABCDEF000000000000000000&path=/*", task.SourceLinkUrl);
-            Assert.True(result);
-        }
-
-        [Fact]
-        public void CustomHosts_Matching5()
-        {
-            var engine = new MockEngine();
-
-            var task = new GetSourceLinkUrl()
-            {
-                BuildEngine = engine,
-                SourceRoot = new MockItem("/src/", KVP("RepositoryUrl", "http://subdomain.tfs.com:100/collection/project/_git/repo"), KVP("SourceControl", "git"), KVP("RevisionId", "0123456789abcdefABCDEF000000000000000000")),
-                Hosts = new[]
-                {
-                    new MockItem("tfs.com", KVP("ContentUrl", "https://domain.com:1")),
-                    new MockItem("tfs.com:123", KVP("ContentUrl", "https://domain.com:2")),
-                    new MockItem("tfs.com:123", KVP("ContentUrl", "https://domain.com:3")),
-                    new MockItem("z.tfs.com", KVP("ContentUrl", "https://domain.com:4")),
-                    new MockItem("z.tfs.com:123", KVP("ContentUrl", "https://domain.com:5")),
-                    new MockItem("z.tfs.com:123", KVP("ContentUrl", "https://domain.com:6")),
-                }
-            };
-
-            bool result = task.Execute();
-            AssertEx.AssertEqualToleratingWhitespaceDifferences("", engine.Log);
-            AssertEx.AreEqual("https://domain.com:1/collection/project/_apis/git/repositories/repo/items?api-version=1.0&versionType=commit&version=0123456789abcdefABCDEF000000000000000000&path=/*", task.SourceLinkUrl);
-            Assert.True(result);
-        }
-
-        [Fact]
-        public void CustomHosts_Matching6()
-        {
-            var engine = new MockEngine();
-
-            var task = new GetSourceLinkUrl()
-            {
-                BuildEngine = engine,
-                SourceRoot = new MockItem("/src/", KVP("RepositoryUrl", "http://subdomain.tfs.com/collection/project/_git/repo"), KVP("SourceControl", "git"), KVP("RevisionId", "0123456789abcdefABCDEF000000000000000000")),
-                RepositoryUrl = "https://tfs.com/collection/project/_git/repo",
-                IsSingleProvider = true,
-                Hosts = new[]
-                {
-                    new MockItem("tfs.com", KVP("ContentUrl", "https://zzz.com")),
-                }
-            };
-
-            bool result = task.Execute();
-            AssertEx.AssertEqualToleratingWhitespaceDifferences("", engine.Log);
-            AssertEx.AreEqual("https://zzz.com/collection/project/_apis/git/repositories/repo/items?api-version=1.0&versionType=commit&version=0123456789abcdefABCDEF000000000000000000&path=/*", task.SourceLinkUrl);
-            Assert.True(result);
-        }
-
-        [Theory]
         [InlineData("", "")]
         [InlineData("", "/")]
         [InlineData("/", "")]
         [InlineData("/", "/")]
-        public void CustomHosts_WithPath1(string s1, string s2)
+        public void BuildSourceLinkUrl(string s1, string s2)
         {
             var engine = new MockEngine();
 
@@ -446,106 +50,6 @@ namespace Microsoft.SourceLink.Tfs.Git.UnitTests
             bool result = task.Execute();
             AssertEx.AssertEqualToleratingWhitespaceDifferences("", engine.Log);
             AssertEx.AreEqual("https://domain.com/x/y/collection/project/_apis/git/repositories/repo/items?api-version=1.0&versionType=commit&version=0123456789abcdefABCDEF000000000000000000&path=/*", task.SourceLinkUrl);
-            Assert.True(result);
-        }
-
-        [Fact]
-        public void CustomHosts_DefaultPortHttp()
-        {
-            var engine = new MockEngine();
-
-            var task = new GetSourceLinkUrl()
-            {
-                BuildEngine = engine,
-                SourceRoot = new MockItem("/src/", KVP("RepositoryUrl", "http://subdomain.tfs.com/collection/project/_git/repo"), KVP("SourceControl", "git"), KVP("RevisionId", "0123456789abcdefABCDEF000000000000000000")),
-                Hosts = new[]
-                {
-                    new MockItem("tfs.com:80", KVP("ContentUrl", "https://domain.com:1")),
-                    new MockItem("tfs.com:443", KVP("ContentUrl", "https://domain.com:2")),
-                    new MockItem("tfs.com:1234", KVP("ContentUrl", "https://domain.com:3")),
-                }
-            };
-
-            bool result = task.Execute();
-            AssertEx.AssertEqualToleratingWhitespaceDifferences("", engine.Log);
-            AssertEx.AreEqual("https://domain.com:1/collection/project/_apis/git/repositories/repo/items?api-version=1.0&versionType=commit&version=0123456789abcdefABCDEF000000000000000000&path=/*", task.SourceLinkUrl);
-            Assert.True(result);
-        }
-
-        [Fact]
-        public void CustomHosts_DefaultPortHttps()
-        {
-            var engine = new MockEngine();
-
-            var task = new GetSourceLinkUrl()
-            {
-                BuildEngine = engine,
-                SourceRoot = new MockItem("/src/", KVP("RepositoryUrl", "https://subdomain.tfs.com/collection/project/_git/repo"), KVP("SourceControl", "git"), KVP("RevisionId", "0123456789abcdefABCDEF000000000000000000")),
-                Hosts = new[]
-                {
-                    new MockItem("tfs.com:80", KVP("ContentUrl", "https://domain.com:1")),
-                    new MockItem("tfs.com:443", KVP("ContentUrl", "https://domain.com:2")),
-                    new MockItem("tfs.com:1234", KVP("ContentUrl", "https://domain.com:3")),
-                }
-            };
-
-            bool result = task.Execute();
-            AssertEx.AssertEqualToleratingWhitespaceDifferences("", engine.Log);
-            AssertEx.AreEqual("https://domain.com:2/collection/project/_apis/git/repositories/repo/items?api-version=1.0&versionType=commit&version=0123456789abcdefABCDEF000000000000000000&path=/*", task.SourceLinkUrl);
-            Assert.True(result);
-        }
-
-        [Fact]
-        public void TrimDotGit()
-        {
-            var engine = new MockEngine();
-
-            var task = new GetSourceLinkUrl()
-            {
-                BuildEngine = engine,
-                SourceRoot = new MockItem("/src/", KVP("RepositoryUrl", "http://mytfs.com/collection/project/_git/repo.git"), KVP("SourceControl", "git"), KVP("RevisionId", "0000000000000000000000000000000000000000")),
-                Hosts = new[] { new MockItem("mytfs.com") }
-            };
-
-            bool result = task.Execute();
-            AssertEx.AssertEqualToleratingWhitespaceDifferences("", engine.Log);
-            AssertEx.AreEqual("https://mytfs.com/collection/project/_apis/git/repositories/repo/items?api-version=1.0&versionType=commit&version=0000000000000000000000000000000000000000&path=/*", task.SourceLinkUrl);
-            Assert.True(result);
-        }
-
-        [Fact]
-        public void TrimmingGitIsCaseSensitive()
-        {
-            var engine = new MockEngine();
-
-            var task = new GetSourceLinkUrl()
-            {
-                BuildEngine = engine,
-                SourceRoot = new MockItem("/src/", KVP("RepositoryUrl", "http://mytfs.com/collection/project/_git/repo.GIT"), KVP("SourceControl", "git"), KVP("RevisionId", "0000000000000000000000000000000000000000")),
-                Hosts = new[] { new MockItem("mytfs.com") }
-            };
-
-            bool result = task.Execute();
-            AssertEx.AssertEqualToleratingWhitespaceDifferences("", engine.Log);
-            AssertEx.AreEqual("https://mytfs.com/collection/project/_apis/git/repositories/repo.GIT/items?api-version=1.0&versionType=commit&version=0000000000000000000000000000000000000000&path=/*", task.SourceLinkUrl);
-            Assert.True(result);
-        }
-
-        [Fact]
-        public void TrimmingGitOnlyWhenSuffix()
-        {
-            var engine = new MockEngine();
-
-            var task = new GetSourceLinkUrl()
-            {
-                BuildEngine = engine,
-                SourceRoot = new MockItem("/src/", KVP("RepositoryUrl", "http://mytfs.com/collection/project/_git/.git"), KVP("SourceControl", "git"), KVP("RevisionId", "0000000000000000000000000000000000000000")),
-                Hosts = new[] { new MockItem("mytfs.com") }
-            };
-
-            bool result = task.Execute();
-            AssertEx.AssertEqualToleratingWhitespaceDifferences("", engine.Log);
-            AssertEx.AreEqual("https://mytfs.com/collection/project/_apis/git/repositories/.git/items?api-version=1.0&versionType=commit&version=0000000000000000000000000000000000000000&path=/*", task.SourceLinkUrl);
             Assert.True(result);
         }
     }
